Chapter 2 task 8

+++ src/main/java/com/walking/intensive/chapter2/task8/Task8.java
@@ -29,7 +29,7 @@ public class Task8 {
             i++;
         }
 
-        return (double) sumEvents / 999999.0;
+        return sumEvents / 999999.0;
     }
 
 }
